% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sctype.R
\name{AnnotateCellTypes}
\alias{AnnotateCellTypes}
\title{Wrapper for sctype by Lanevski et al.}
\usage{
AnnotateCellTypes(
  object,
  genes.list = NULL,
  assay = "RNA",
  is.scaled = TRUE,
  tissue = NULL
)
}
\arguments{
\item{object}{An object of class Seurat}

\item{genes.list}{A named list}

\item{assay}{The assay from Seurat to use. Default is RNA}

\item{is.scaled}{Is the input data scaled. Should be scaled}

\item{tissue}{Tissue type supplied if gene list of markers is not}
}
\description{
Wrapper for sctype by Lanevski et al.
}
