% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_functions.R
\name{apply_DoubletFinder}
\alias{apply_DoubletFinder}
\title{Run DoubletFinder on an Seurat object}
\usage{
apply_DoubletFinder(
  seurat_object,
  npcs_PCA = 50,
  pN_fixed = 0.25,
  pK_max = NULL,
  doublet_formation_rate = 0.075,
  adjust_nExp = FALSE,
  doublet_cluster_tresh = NULL,
  cluster_column = "seurat_clusters",
  return_seurat = TRUE
)
}
\arguments{
\item{seurat_object}{seurat object}

\item{npcs_PCA}{how many PCs to use (DoubletFinder wil run its own PCA!)}

\item{pN_fixed}{pN to pass to DoubletFidner}

\item{pK_max}{pK will be estimated, this parameter allows to se a maximum cap}

\item{doublet_formation_rate}{how many doublets are expected. Deafults to 0.075}

\item{adjust_nExp}{whether to adjust number of doublets with modelHomotypic}

\item{doublet_cluster_tresh}{defaults to NULL. if not NULL: Clusters with more than doublet_cluster_tresh pct of cells will completely be considered Doublets}

\item{cluster_column}{metadata column with clusters}

\item{return_seurat}{whether to return seurat with 'Doublet' column}
}
\value{
seurat with 'Doublet' column or named vector
}
\description{
Wrapper around Seurat::FindClusters
}
\details{
see also: https://github.com/chris-mcginnis-ucsf/DoubletFinder
}
