% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{CalculateMultScore}
\alias{CalculateMultScore}
\title{Calculate a gene set score for each cell in a Seurat object based on teh product (if one gene is missing the score will be zero)}
\usage{
CalculateMultScore(
  seurat_object,
  features,
  cells = NULL,
  allowed_zeros = 0,
  allowed_zero_pct = 0
)
}
\arguments{
\item{seurat_object}{seurat object}

\item{features}{which genes to calculate score on}

\item{cells}{subset to tehses cells (defaults to NULL which emans all)}

\item{allowed_zeros}{how many zeros (integer) are allowed (will be ignored when calculating product)}

\item{allowed_zeros_pct}{how many zeros (pct of number of features) are allowed. Number between 0 and 1. Rounded with floor. If larger than 0 this will overwrite allowed_zeros}
}
\value{
a numeric vector with scores per cell in seurat_object
}
\description{
Useful for highlighting all cells that express two or three genes together.
Scores for longer gene sets will likely be zero, but the allowed_zeros(pct) argument can be used to allow a certain number of zeros (and calculation of the product only on the other non-zero genes).
}
