% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_variable_features.R
\name{CleanVarGenes}
\alias{CleanVarGenes}
\title{Remove Mitochondrial and Ribosomal Genes from Variable Features in Seurat}
\usage{
CleanVarGenes(object, nHVG = 2000)
}
\arguments{
\item{object}{An object of class Seurat}

\item{nHVG}{Number of variable features to clean}
}
\value{
A Seurat object with variable features that do not contain mito or rb genes
}
\description{
Remove Mitochondrial and Ribosomal Genes from Variable Features in Seurat
}
