% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{closed_breaks_log}
\alias{closed_breaks_log}
\title{Closed Breaks for log scales}
\usage{
closed_breaks_log(lims, n = 5, base = 10)
}
\arguments{
\item{lims}{Vector with lower and upper limits of the data that you want
breaks for.}

\item{n}{Number of breaks to aim for. The actual number will be what the
algorithm can accomodate, see \code{?scales::breaks_log()}. Default: 5}

\item{base}{The base of logarithm to use. Default: 10}
}
\value{
Numeric vector with breaks.
}
\description{
Same algorithm as in \link[scales]{breaks_log}, but
forces inclusion of upper and lower limits (displaying the closed interval).
Including boundaries is particularly useful for ggplot2's color/fill, as it
emphasizes the meaning of maximal/minimal color intensities, see examples.
}
\details{
Code adapted from scales::breaks_log, original code is
[licensed by Hadley Wickham](https://cran.r-project.org/web/packages/scales/LICENSE),
changes are licensed by Felix Frauhamer.
Original license is MIT:

\itemize{
  \item YEAR: 2010-2016
  \item COPYRIGHT HOLDER: Hadley Wickham}


The scUtils pacakges uses this function to color by gene expression,
where the maximal expression gives valuable
intuition for a gene's overall expression strength. For x- or y-axis, I still
recommend breaks_log from the scales package.
}
\examples{
\dontrun{
if(interactive()){
# closed breaks include maximum, breaks_log do not:
closed_breaks_log(lims = c(.01, 977.1))
scales::breaks_log()(c(.01, 977.1))
# closed_breaks_log intends to replace outer breaks with actual limits:
round(scales::breaks_log(base=2)(c(.01, 977.1)), 3)
closed_breaks_log(lims = c(.01, 977.1), base=2)
 }
}
}
\seealso{
\code{\link[assertthat]{assert_that}}
}
