% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{conserved_correlations}
\alias{conserved_correlations}
\title{Calculate correlation between features and all other genes in a seurat object per dataset}
\usage{
conserved_correlations(
  seurat_object,
  slot = "data",
  assay = "RNA",
  features,
  conserve_col = "Dataset",
  min_cells = 20,
  method = "correlation"
)
}
\arguments{
\item{seurat_object}{seurat object}

\item{slot}{which slot to pull data from}

\item{assay}{which assay}

\item{features}{which genes to calculate correlation with}

\item{conserve_col}{column name from seurat_object meta.data}

\item{min_cells}{show many cells at least to calculate correlation for group}

\item{method}{valid method passed to proxyC::simil}
}
\value{
a list of matrices (one per gene in features) with rows as genes and columns as categories of conserve_col
}
\description{
If batch effects occur (or per cluster correlations are desired), one correlation value per group can be calculated.
Uses the proxyC library, any valid method can be provided
}
