% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion_functions.R
\name{convert_species_seurat}
\alias{convert_species_seurat}
\title{Converts the gene annotation in a seurat object from mouse to human or vice versa}
\usage{
convert_species_seurat(
  seurat_object,
  add_as_assay = FALSE,
  assay = "RNA",
  host = "https://feb2021.archive.ensembl.org",
  verbose = TRUE
)
}
\arguments{
\item{seurat_object}{Seurat Object to convert}

\item{add_as_assay}{boolean: Whether to add the data as a second assay named 'assay'_'species'}

\item{assay}{The seurat assay to convert. defaults to RNA}

\item{host}{the ensembl host . Relevant for the genome version to use. Set to "https://www.ensembl.org" for latest}

\item{verbose}{boolean whether to print messages}
}
\value{
species converted seurat object (see also add_as_assay)
}
\description{
At the moment works only with gene names (symbols). Not Ids.
}
