% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvscore.R
\name{CVScore}
\alias{CVScore}
\title{CV Score Wrapper for Seurat object as described by Lakkis et. al (2021)}
\usage{
CVScore(
  object,
  groupby = "seurat_clusters",
  batch = "tumorID",
  weighted = TRUE
)
}
\arguments{
\item{object}{An object of class Seurat}

\item{groupby}{Clustering metadata variable}

\item{batch}{Batch metadata variable}

\item{weighted}{Boolean to determine if weighted score should be calculated for all clusters}
}
\value{
A vector of scores
}
\description{
CV Score Wrapper for Seurat object as described by Lakkis et. al (2021)
}
