% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_functions.R
\name{determine_cluster_resolution}
\alias{determine_cluster_resolution}
\title{Find resolution closest to target_cluster_number}
\usage{
determine_cluster_resolution(
  seurat_object,
  target_cluster_number,
  resolutions = c(0.5, 0.75, 1, 1.5, 2:10),
  min_cells = 5,
  graph_name = "RNA_snn",
  cluster_col_name = "seurat_clusters",
  return_seurat = TRUE,
  seed = 1234,
  ...
)
}
\arguments{
\item{seurat_object}{seurat object}

\item{target_cluster_number}{integer of desired unique clusters}

\item{resolutions}{resolutions to calculate}

\item{min_cells}{min cells for a cluster to count (default: 5)}

\item{graph_name}{graph_name in seurat_object (default: 'RNA_snn')}

\item{cluster_col_name}{name for clumn in seurat object that will contain the clusters from the best resolution. (Defaults to 'seurat_clusters')}

\item{return_seurat}{return cluster resolution name or seurat object with column cluster_col_name}

\item{seed}{random seed}

\item{...}{other parameters passed on to Seurat::FindClusters}
}
\value{
seurat_object with variable features in misc$var_features
}
\description{
Wrapper around Seurat::FindClusters
}
