% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{downsample_balanced_iterative}
\alias{downsample_balanced_iterative}
\title{Helper function for a balanced downsampling procedure.}
\usage{
downsample_balanced_iterative(
  metadata,
  target_sample_size,
  predictor_var,
  stepsize = 100,
  global_seed = 1234
)
}
\arguments{
\item{metadata}{metadata with a Cell_ID column and predictor_var}

\item{target_sample_size}{the target number of cells after downsampling}

\item{predictor_var}{name of metadata column that will be predicted (celltype)}

\item{stepsize}{stepsize for iterative downsampling (smaller is more accurate but slightly slower)}

\item{global_seed}{seed used for random forests}
}
\value{
downsampled metadata
}
\description{
The goal is not to have an equal amount of all cells but instead to sample to a target number with the larger classes being downsampled more strongly.
}
