% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicates.R
\name{dupes}
\alias{dupes}
\title{Return rows with duplicate values}
\usage{
dupes(x, idCols)
}
\arguments{
\item{x}{a dataframe}

\item{idCols}{character vector of column names to check for duplicates.}
}
\value{
a dataframe containing only rows with duplicated values in the ID columns.
}
\description{
Return all rows with duplicated values in the ID columns.
}
