% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicates.R
\name{duplicated2}
\alias{duplicated2}
\title{Determine all duplicated elements.
return TRUE for all duplicated entries, including the first one of each set.}
\usage{
duplicated2(x)
}
\arguments{
\item{x}{a vector or a data frame or an array or NULL}
}
\value{
A logical vector (TRUE for duplicates and FALSE for non duplicates)
}
\description{
'duplicated2' returns a logical vector: TRUE for all the
duplicated elements, and FALSE for unique elements. Whereas duplicated(), a base function, does not
include the first entry of the duplicated set.
}
