% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect_in.R
\name{expect_in}
\alias{expect_in}
\title{Expect in}
\usage{
expect_in(object, expected)
}
\arguments{
\item{object}{actual output set after running the function}

\item{expected}{expected output set that will include the actual output set.
but may or may not entirely match the length and order of actual output set.}
}
\value{
object actual output set

expected expected output set
}
\description{
testthat, this function is a modified version of
testthat::expect_setequal. This function will check if actual set of the output
is included or is present in the expected set.
}
