% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbind.R
\name{fbind}
\alias{fbind}
\title{Binds two factors, here sentences}
\usage{
fbind(a, b)
}
\arguments{
\item{a}{If a is defined as: (a <- factor(c("South", "Carolina","water", "use", "data")))}

\item{b}{If b is defined as: (b <- factor(c("and", "package", "building", "exercise")))
c (a,b)
factor(c(as.character(a), as.character(b)))}
}
\value{
: If fbind(a,b) is entered and run then it will join or bind those two sentences
}
\description{
Binds two factors, here sentences
}
