% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_functions.R
\name{feature_statistics}
\alias{feature_statistics}
\title{Calculate mean and pct of features (genes) in seurat objects in a flexible manner}
\usage{
feature_statistics(
  object,
  features = NULL,
  cells = NULL,
  assay = "RNA",
  slot = "data",
  thresh.min = 0,
  digits = 3,
  result_colnames = c("gene", "mean", "pct")
)
}
\arguments{
\item{object}{seurat object}

\item{features}{which features to check. Will ignore non-existent features. Defaults to NULL which means all features.}

\item{cells}{which cells to subset the object to before calculating statistics. Will ignore non-existent cell names. Defaults to NULL which means all cells.}

\item{assay}{seurat assay to pull data from. Defaults to 'RNA'}

\item{slot}{slot in assay to use. Defaults to 'data'}

\item{thresh.min}{min expression of a feature to be considered for percentage calculation}

\item{digits}{how many digits for rounding the results}

\item{result_colnames}{how to name the columns in the results. Defaults to c("gene", "mean", "pct")}
}
\value{
data frame with one row per feature and 3 columns
}
\description{
Similar to Seurat's calculation of summary statistics for FindMarkers
}
