% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureTest.R
\name{FeatureTest}
\alias{FeatureTest}
\title{FeatureTest}
\usage{
FeatureTest(
  SeuratObject,
  features,
  assay = "RNA",
  layers = NULL,
  thresholds = "mad",
  operators = "both",
  nmad = 3,
  percentile = 1
)
}
\arguments{
\item{SeuratObject}{A Seurat object containing the data.}

\item{features}{The features (gene or metadata variable) to compute the threshold for.
You can provide the same feature multiple times with different thresholds and operators
to compute multiple tests for the same feature. FeatureTest will automatically handle
the repeated features and compute the results for each test separately, and the final
output will be TRUE if all tests for the same feature pass, and FALSE if any test fails.}

\item{assay}{The assay to use for fetching the feature data.}

\item{layers}{The layer type to use for fetching the feature data (counts, data...).
If NULL, the default layer (meta.data) is used. (NAs, emtpy strings, and "null" are treated as NULL).
In the case of a split SeuratObject, the function will automatically compute the threshold for each layer separately
respecting the data structure.}

\item{thresholds}{A list of thresholds to use for computing the pass/fail results.
To use Median Absolute Deviation (MAD) or percentile method, use mad or percentile respectively.}

\item{operators}{A list of operators to use for computing the pass/fail results.
Valid operators are "upper", "lower", or "both" in the case of MAD or percentile method.
For numeric thresholds, input the desired operators as strings. Valid operators: ">", "<", ">=", "<=", "==", "!=".
Use params nmad and percentile to set the number of MADs or the quantile to consider as outliers.}

\item{nmad}{The number of MADs to use to compute the threshold.}

\item{percentile}{The quantile to consider as outliers. It must be an integer between 1 and 99.
Default is 1 and computes the top and bottom 1% of the distribution.}
}
\value{
SeuratObject A Seurat object with the computed threshold and pass/fail results added to the meta.data slot.
}
\description{
A helper function to compute the threshold and pass/fail results for a given features.
FeatureTest work by fetching the feature data from the Seurat object,
computing the threshold based on the specified method (MAD,percentile, or user's threshold),
and determining whether each cell passes or fails the threshold test.
The results are then added to the Seurat object's metadata.
For each feature, the user must imput a threshold and a "operator". Operators determine the
comparison that it will be performed between the feature values and the threshold.
If multiple features are given, but only one threshold and operator are provided,
the same threshold and operator will be applied to all features.
}
\examples{
\dontrun{

library(Seurat)
library(dplyr)
library(data.table)
library(purrr)

seurat_object <- FeatureTest(
  SeuratObject = seurat_object,
  features = c("Gene1", "Gene2", "percent.mt"),
  assay = "RNA",
  layers = c("couts", "data", NA),
  thresholds = "mad",
  operators = "both",
  nmad = 3,
  percentile = 1
)
}

}
