% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{FilterVariableFeatures}
\alias{FilterVariableFeatures}
\title{FilterVariableFeatures}
\usage{
FilterVariableFeatures(
  SeuratObject,
  pattern = NULL,
  assay = "RNA",
  verbose = TRUE
)
}
\arguments{
\item{SeuratObject}{A Seurat object containing variable features.}

\item{pattern}{A character vector of regular expression patterns to filter variable features.
If NULL, default patterns will be used to remove TCR/BCR and IG genes,
HLA genes, mitochondrial and ribosomal genes, and non-coding and antisense genes.}

\item{assay}{The assay to filter variable features from. Defaults to "RNA".}

\item{verbose}{Logical. If TRUE, prints the number of features removed and remaining.}
}
\value{
A Seurat object with filtered variable features.
}
\description{
Filters the variable features of a Seurat object to remove TCR/BCR and IG genes,
mitochondrial and reibosomal genes, and non-coding and antisense genes.
}
