% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{find_children}
\alias{find_children}
\title{Recursively walk through edgelist and return children of specified node(s).}
\usage{
find_children(nodes, edges)
}
\arguments{
\item{nodes}{vector of node(s) present in edges for which all children will be determined}

\item{edges}{edgelist with two columns: 'to' and 'from'}
}
\value{
vector with children ids
}
\description{
Recursively walk through edgelist and return children of specified node(s).
}
