% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans_seurat.R
\name{FindClustersKmeans}
\alias{FindClustersKmeans}
\title{Wrapper function to run Kmeans Clustering on a Seurat object}
\usage{
FindClustersKmeans(object, k = 2, reduction = "pca", ...)
}
\arguments{
\item{object}{An object of class Seurat}

\item{k}{Parameter for number of clusters in kmeans}

\item{reduction}{Dimension reduction to use from Seurat object for clustering}

\item{...}{Any other parameters for Base R kmeans function}
}
\value{
A Seurat object with kmean clusters added as metadata
}
\description{
Wrapper function to run Kmeans Clustering on a Seurat object
}
