% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_gene_expr_functions.R
\name{FindDEG_nebula}
\alias{FindDEG_nebula}
\title{Run nebula on an seurat object with condition(s) and multiple subjects}
\usage{
FindDEG_nebula(
  seurat_object,
  cluster_variable = "seurat_clusters",
  sample_variable = "Sample_ID",
  primary_variable = "Condition",
  other_variables = c(),
  features = NULL,
  nCounts = NULL,
  min_cells = 5,
  min_pct_valid_samples = 0.5,
  assay = "RNA",
  padjust_method = "fdr",
  return_full_result = TRUE,
  numCores = 1,
  verbose = TRUE
)
}
\arguments{
\item{seurat_object}{a seurat object with all required metadata}

\item{cluster_variable}{column name in seurat meta.data with the clusters}

\item{sample_variable}{column name in seurat meta.data with the subject or sample id}

\item{primary_variable}{column name in seurat meta.data with the main variable to compare (condition, treatment etc.)}

\item{other_variables}{column names in seurat meta.data with other relevant variables to include in nebula model}

\item{features}{subset to these features (speeds up!). NULL means use all features. Additionally nebula's default setting for lowly expressed gene removal is applied.}

\item{nCounts}{column name in seurat meta.data with the scaling factor ("offset" in nebula). Can be NULL if all features are used. Then the scaling factor is the total counts number.}

\item{min_cells}{number of cells (in currently evaluated cluster) for a sample to be valid}

\item{min_pct_valid_samples}{how many valid samples have to be in cluster in order to include this cluster. 0.5 means 50\% of samples must have at least min_cells in the cluster. Set this to 1 to requires all samples to have min_cells.}

\item{assay}{"RNA" or other assay from seurat object}

\item{padjust_method}{defaults to "fdr" . directly passed to stats::p.adjust}

\item{return_full_result}{return only marker table or list with all nebula outputs}

\item{numCores}{number of cores used by foreach}

\item{verbose}{whether to print verbose messages}
}
\value{
dataframe with DEGs or list with dataframe and list off full nebula results depending on return_full_result
}
\description{
Wrapper around run_nebula to execute it for all clusters.
}
\details{
requires nebula and Rfast to be installed : https://github.com/lhe17/nebula
Requires doParallel and foreach to be installed
}
