% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_gene_expr_functions.R
\name{FoldChange_seurat}
\alias{FoldChange_seurat}
\title{Calculate general FC and pcts between two cell groups}
\usage{
FoldChange_seurat(object, cells.1, cells.2, features = NULL)
}
\arguments{
\item{cells.1}{Vector of cell names belonging to group 1}

\item{cells.2}{Vector of cell names belonging to group 2}

\item{features}{Features to calculate fold change for.
If NULL, use all features}
}
\description{
https://github.com/satijalab/seurat/blob/master/R/differential_expression.R
}
