% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{gene_pct_cluster}
\alias{gene_pct_cluster}
\title{Get per cluster gene pct}
\usage{
gene_pct_cluster(
  seurat_object,
  genes,
  col_name,
  min_expression = 0,
  return_long = FALSE
)
}
\arguments{
\item{seurat_object}{seurat object}

\item{genes}{which genes}

\item{col_name}{metadata_column}

\item{min_expression}{min expressio. defaults to 0}

\item{return_long}{return data frame in lonf format}
}
\value{
matrix or long dataframe
}
\description{
Get per cluster gene pct
}
