% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion_functions.R
\name{get_mouse_human_genes_conversion}
\alias{get_mouse_human_genes_conversion}
\title{Match samples between two metadata samples based on a third column (Barcodes)}
\usage{
get_mouse_human_genes_conversion(host = "https://feb2021.archive.ensembl.org")
}
\arguments{
\item{host}{defaults to https://feb2021.archive.ensembl.org Can also use https://nov2020.archive.ensembl.org or "https://www.ensembl.org" for latest}
}
\value{
a data.frame with columns mouse_gene and human_gene
}
\description{
Requires a similarly named column with corresponding barcodes in both tables
}
\details{
hsapiens_homolog_perc_id:  \%id. target Human gene identical to query gene
hsapiens_homolog_perc_id_r1: \%id. query gene identical to target Human gene
}
