% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nin.R
\name{\%nin\%}
\alias{\%nin\%}
\title{Not in}
\usage{
x \%nin\% y
}
\arguments{
\item{x, y}{vectors or NULL.
Long vectors not supported for y.}
}
\value{
A logical vector of the same length as x.
}
\description{
Binary operator which returns a logical vector indicating
if there is no match between contents of x and y.
}
