% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_functions.R
\name{identify_variable_features}
\alias{identify_variable_features}
\title{Find HVGs with vst.log per Batch (or other Seurat FindVariableFeatures method)}
\usage{
identify_variable_features(
  seurat_object,
  n_hvgs_sizes = 1000,
  batch_var,
  assay_name = "RNA",
  method = "vst",
  ignore_genes_regex = NULL,
  ignore_genes_vector = NULL,
  returnSeurat = TRUE,
  seed = 123
)
}
\arguments{
\item{seurat_object}{seurat object}

\item{n_hvgs_sizes}{integer or integer vector of desired feature set sizes}

\item{batch_var}{var to split by}

\item{assay_name}{seurat object assay}

\item{method}{which HVG method : selection.method from Seurat::FindVariableFeatures}

\item{ignore_genes_regex}{a regex of genes to filter from HVGs}

\item{ignore_genes_vector}{a character vector of genes to remove from HVGs}

\item{returnSeurat}{return seurat}

\item{seed}{random seed}
}
\value{
seurat_object with variable features in misc$var_features
}
\description{
Find HVGs with vst.log per Batch (or other Seurat FindVariableFeatures method)
}
