% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{infer_sex}
\alias{infer_sex}
\title{Infer sex per sample based on Xist expression in cells}
\usage{
infer_sex(
  seurat_object,
  sample_column,
  min_xist_female = 0.7,
  max_xist_male = 0.1
)
}
\arguments{
\item{seurat_object}{a seurat object}

\item{sample_column}{column with sample annotation}

\item{min_xist_female}{min pct of cells in sample that are xist positive to declare female}

\item{max_xist_male}{max pct of cells in sample that are xist positive to declare male}
}
\value{
a vector with either F,M or U per cell (but based on the samples)
}
\description{
The goal is not to have an equal amount of all cells but instead to sample to a target number with the larger classes being downsampled more strongly.
}
