% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect_in.R
\name{is_vector}
\alias{is_vector}
\title{Check for vector}
\usage{
is_vector(x)
}
\arguments{
\item{x}{list or atomic}
}
\description{
used within the expect_in function to check if input is a list
or atomic
}
