% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{is_wholenumber}
\alias{is_wholenumber}
\title{Check if number(s) is/are integers. In contrast to is.integer, is_wholenumber
does not check the class but accepts all numbers that are integers with reasonable
precision.}
\usage{
is_wholenumber(x, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{x}{Number to test}

\item{tol}{tolerance for testing}
}
\description{
Check if number(s) is/are integers. In contrast to is.integer, is_wholenumber
does not check the class but accepts all numbers that are integers with reasonable
precision.
}
