% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect_in.R
\name{locations}
\alias{locations}
\title{Detect location}
\usage{
locations(i)
}
\arguments{
\item{i}{element from actual set}
}
\description{
used within the expect_in function to check which element in actual
set is missing from expected set
}
