% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{match_sample_names}
\alias{match_sample_names}
\title{Match samples between two metadata samples based on a third column (Barcodes)}
\usage{
match_sample_names(
  table_A,
  table_B,
  sample_col_A,
  sample_col_B,
  barcode_col = "Barcode",
  min_pct = 0.5,
  silent = FALSE
)
}
\arguments{
\item{table_A}{a data.frame (metadata from single cell data)}

\item{table_B}{a second data.frame (metadata from single cell data)}

\item{sample_col_A}{column name with sample names in A}

\item{sample_col_B}{column names with sample names in B}

\item{barcode_col}{column with barcodes (in btoh tables}

\item{min_pct}{minimum of matching barcodes to set as sample}

\item{silent}{messages}
}
\value{
a data.frame ith matched samples (NA if no matchin sample in B for A). Will drop samples from B that cannot be matched to A.
}
\description{
Requires a similarly named column with corresponding barcodes in both tables
}
