% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/narm.R
\name{na.rm}
\alias{na.rm}
\title{Remove Missing Values}
\usage{
na.rm(x)
}
\arguments{
\item{x}{a vector.}
}
\value{
A vector the same as x but with missing values removed.
}
\description{
Remove missing values (NA) from a vector
}
