% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{rasterize_ggplot}
\alias{rasterize_ggplot}
\title{Rasterize ggplot}
\usage{
rasterize_ggplot(
  plot,
  pixel_raster = 1024,
  pixel_raster_y = NULL,
  interpolate = FALSE,
  pointsize = 1
)
}
\arguments{
\item{plot}{ggplot2 or patchwork object. For example output of Seurat's DimPlot or FeaturePlot}

\item{pixel_raster}{integer: number of pixels passed to pixels argument (x and optionall y) from scattermore::geom_scattermore. Defaults to 1024}

\item{pixel_raster_y}{pixel_raster_y: to use a different y pixels than x. Defaults to NULL which will use the value from pixel_raster in x and y}

\item{interpolate}{whether to linearly interpolate the image. Defaults to FALSE}

\item{pointsize}{Radius of rasterized point from scattermore. defaults to 1}
}
\value{
ggplot object with GeomPoint layers changed to GeomScattermore with rastrized version
}
\description{
Makes use of scattermore::geom_scattermore to substitute GeomPoint layer in ggplot objects with an equivalent rasterized version.
Allows to manually modify the resolution of rasterized plots from Seurat plots!
}
