% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_functions.R
\name{Read10xFormat}
\alias{Read10xFormat}
\title{Read10xFormat}
\usage{
Read10xFormat(mtx, cells, features, cell.column = 1, feature.column = 2)
}
\arguments{
\item{mtx}{Name of the mtx file}

\item{cells}{Name of the cells file}

\item{features}{Name of the features file}

\item{cell.column}{Specify which column of cells file to use for cell names; default is 1}

\item{feature.column}{Specify which column of features files to use for feature/gene names; default is 2}
}
\value{
dgCMatrix with row and col names
}
\description{
Similar to Seurat's ReadMtx but with less error handling and options
Sometimes ReadMtx does not work --> then use this
}
