% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_functions.R
\name{ReadDGEFormat}
\alias{ReadDGEFormat}
\title{ReadDGEFormat}
\usage{
ReadDGEFormat(dge, feature.column = 1)
}
\arguments{
\item{dge}{Name of the dge file}

\item{feature.column}{Specify which column of the dge matrix will be used as rownames (cell column start from feature.column+1)}
}
\value{
dgCMatrix with row and col names
}
\description{
Read dge file
}
