% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicates.R
\name{removeIncompleteDupes}
\alias{removeIncompleteDupes}
\title{Remove incomplete duplicate rows.}
\usage{
removeIncompleteDupes(data, id, x, y)
}
\arguments{
\item{data}{A data frame}

\item{id}{The id column, which should be unique but isn't.}

\item{x}{A column of interest, perhaps with some values missing (NA).}

\item{y}{Another column of interest, perhaps with some values missing (NA).}
}
\description{
This function takes a table, an id column, and two data columns.
}
