% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_gene_expr_functions.R
\name{run_nebula}
\alias{run_nebula}
\title{Apply nebula on input data}
\usage{
run_nebula(
  counts,
  metadata,
  formula_string = "~ treatment",
  subject_id = "Subject_ID",
  cell_id = "Cell_ID",
  cell_size = NULL
)
}
\arguments{
\item{counts}{count matrix Dgc}

\item{metadata}{...}

\item{formula_string}{...}

\item{subject_id}{...}

\item{cell_id}{...}

\item{cell_size}{...}
}
\value{
nebula output list
}
\description{
Wrapper around model.matrix, nebula::group_cell, nebula::nebula
}
\details{
requires nebula and Rfast to be installed : https://github.com/lhe17/nebula
}
