% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scvi.R
\name{RunSCVI}
\alias{RunSCVI}
\title{Wrapper for using scVI for list of Seurat objects to correct for batch effects}
\usage{
RunSCVI(
  object.list,
  batch.key = NULL,
  features = 2000,
  assay = "RNA",
  merge.data = TRUE,
  vars.to.regress = NULL
)
}
\arguments{
\item{object.list}{A list of Seurat objects}

\item{batch.key}{Metadata variable or variables to indicate batch}

\item{features}{Either a list of variable features or numeric to calculate number of variable features used for integration}

\item{assay}{Assay to calculate scVI. Default is RNA from Seurat}

\item{merge.data}{If list of objects are normalized using the same method, values are carried over when merging objects}

\item{vars.to.regress}{List of features to regress such as cell cycle, if calculated on each Seurat object}
}
\value{
Integrated Seurat object using scVI
}
\description{
Wrapper for using scVI for list of Seurat objects to correct for batch effects
}
