% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scGSEAmarkers.R
\name{scGSEAmarkers}
\alias{scGSEAmarkers}
\title{scGSEAmarkers}
\usage{
scGSEAmarkers(
  cluster_markers,
  reference_markers,
  padj.threshold = 1e-06,
  only.pos = TRUE,
  workers = 4
)
}
\arguments{
\item{cluster_markers}{A data.frame containing the DE markers for multiple clusters
from a SC experiment (i.e. a SeuratObject). Expects the output from Seurat::FindAllMarkers.}

\item{reference_markers}{A GSEA database loaded for fgsea input (see fgsea documentation).}

\item{padj.threshold}{Threshold for p-adj. Default is 1e-6.}

\item{only.pos}{Returns only positively enriched markers (NES > 0). Default is true.}

\item{workers}{Number of cores for parallel computing. Default is 4.}
}
\value{
A list containing the GSEA for each cluster.
}
\description{
Calculates GSEA for all the clusters in a SeruatObject. Given the output from Seurat::FindAllMarkers(), calculate the GSEA for all clusters given a GSEA database. For example, use MSigDB C8 dataset to calculate enrichment for cell identity markers. This is intended to be a helper for manual annotation of the clusters.
}
\examples{
\dontrun{
results <- scGSEAmarkers(cluster_markers, reference_markers)
}
}
