% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{SelectPCs}
\alias{SelectPCs}
\title{SelectPCs}
\usage{
SelectPCs(SeuratObject, seed = 42L)
}
\arguments{
\item{SeuratObject}{A Seurat object.}
}
\value{
Numerical A vector of selected principal components.
}
\description{
Select the number of principal components to use for downstream analysis.
The function uses the intrinsicDimension::maxLikGlobalDimEst to estimate the intrinsic
dimensionality of the PCA embeddings in the Seurat object. It handles large datasets by chunking
the data to avoid memory issues and overflow errors. If the intrinsic dimension cannot be
estimated, it falls back to a pointwise estimation method and takes the median of the valid
estimates as a robust global estimate.
}
