% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_functions.R
\name{seurat_recipe}
\alias{seurat_recipe}
\title{Apply a standard seurat workflow to a seurat object}
\usage{
seurat_recipe(
  seurat_object,
  assay = "RNA",
  nfeatures_vst = 2000,
  clean_hvg = FALSE,
  normalize_data = TRUE,
  npcs_PCA = 50,
  findClusters = F,
  calcUMAP = TRUE,
  clusterRes = 1,
  key = "pca",
  seed = 123
)
}
\arguments{
\item{seurat_object}{seurat object}

\item{assay}{which assay}

\item{nfeatures_vst}{number of features for HVG detection}

\item{clean_hvg}{clean up HVGs with basic regex}

\item{normalize_data}{normalize data with standard logNorm}

\item{npcs_PCA}{how many pcs to save}

\item{findClusters}{run neighbor and cluster detection with seurat standard}

\item{calcUMAP}{calc standard seurat UMAP}

\item{clusterRes}{resolution for clustering}

\item{key}{name/key for dimred (will name PCA like this and then use for all following steps)}

\item{seed}{random seed}
}
\value{
processed seurat_object
}
\description{
Apply a standard seurat workflow to a seurat object
}
