% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soup_correct.R
\name{SoupCorrect}
\alias{SoupCorrect}
\title{Wrapper for Ambient RNA removal using SoupX}
\usage{
SoupCorrect(raw.matrix, filt.matrix, contamination_rate = NULL)
}
\arguments{
\item{raw.matrix}{Raw expression matrix from 10x Chromium}

\item{filt.matrix}{Filtered expression matrix 10x Chromium}

\item{contamination_rate}{Contamination rate to correct for soup}
}
\value{
Adjusted count matrix to be used downstream
}
\description{
Wrapper for Ambient RNA removal using SoupX
}
