% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariableFeaturePlot2.R
\name{VariableFeaturePlot2}
\alias{VariableFeaturePlot2}
\title{VariableFeaturePlot2}
\usage{
VariableFeaturePlot2(
  object,
  cols = c("black", "red"),
  pt.size = 1,
  log = NULL,
  selection.method = NULL,
  assay = NULL,
  raster = NULL,
  raster.dpi = c(512, 512),
  hvf = NULL,
  label = FALSE,
  n.top.hvf = 10,
  custom.label = NULL,
  repel = TRUE
)
}
\arguments{
\item{object}{Seurat object}

\item{cols}{Colors to specify non-variable/variable status}

\item{pt.size}{Size of the points on the plot}

\item{log}{Plot the x-axis in log scale}

\item{selection.method}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}}

\item{assay}{Assay to pull variable features from}

\item{raster}{Convert points to raster format, default is \code{NULL}
which will automatically use raster if the number of points plotted is greater than
100,000}

\item{raster.dpi}{Pixel resolution for rasterized plots, passed to geom_scattermore().
Default is c(512, 512).}

\item{hvf}{The list of highly variable features to use for plotting. If \code{NULL}, the function
will use the variable features stored in the Seurat object.}

\item{label}{Logical, whether to label the top variable features on the plot. Default is \code
{FALSE}.}

\item{n.top.hvf}{Number of top variable features to label if \code{label} is \code{TRUE}.
Default is 10.}

\item{custom.label}{A character vector of specific features to label on the plot. If provided,
these features will be labeled instead of the top variable features. Default is \code{NULL}.}

\item{repel}{Logical, whether to use ggrepel for labeling points. Default is \code{TRUE}.}
}
\value{
A ggplot object
}
\description{
This function is a modified version of Seurat's VariableFeaturePlot.
It allows users to specify a custom list of highly variable features (hvf) for plotting,
rather than relying solely on the variable features stored in the Seurat object. It also forces
the use of the output of SeuratObject::VariableFeatures() to highligh the high variable features
when the hvf parameter is not provided. This can be useful for visualizing specific sets of
features, for example, those which have been filtered based on certain criteria.
}
\details{
This function is designed to provide flexibility in visualizing variable features in a Seurat
object. By allowing users to specify a custom list of highly variable features, it enables the
visualization of specific features of interest, rather than being limited to the variable
features identified by Seurat's default methods.
}
\examples{
\dontrun{
  data("pbmc_small")
  pbmc_small <- data("pbmc_small") |>
    Seurat::NormalizeData() |>
    Seurat::FindVariableFeatures() |>
    SCutils::FilterVariableFeatures() |>
    Seurat::ScaleData()
  hvf <- SeuratObject::VariableFeatures(pbmc_small)
  # Plot a unique variable feature plot for all layers but highligh
  # the consensus variable features across all layers instead of the r
  # recalculated variable features for each layer.
  pbmc_small |>
    JoinLayers() |>
    Seurat::FindVariableFeatures() |>
    SCutils::VariableFeaturePlot2(hvf = hvf)
}

}
