% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{writeList_to_JSON}
\alias{writeList_to_JSON}
\title{Write a list of strings (will cast to string) to a json to have a simple conversion to python.}
\usage{
writeList_to_JSON(list_with_rows, filename)
}
\arguments{
\item{list_with_rows}{The list that will be written}

\item{filename}{the destination without ending!}
}
\value{
nothing
}
\description{
Write a list of strings (will cast to string) to a json to have a simple conversion to python.
}
