% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BatchOpenH5.R
\name{BatchOpenH5}
\alias{BatchOpenH5}
\title{Batch-open multiple .h5ad files, generates the BP matrices and simple metadata, and converts features to gene symbols}
\usage{
BatchOpenH5(
  files,
  BP.data.dir = NULL,
  platform = "10X",
  use.names = TRUE,
  ensembl.to.symbol = FALSE,
  species = "human",
  generate.metadata = FALSE,
  mc.cores = length(files)
)
}
\arguments{
\item{files}{Character vector. Filepaths (including the .h5* extension) to process.}

\item{BP.data.dir}{Character. Directory where the BP matrices will be saved
(without the `"_BP"` suffix). If `NULL`, uses the first file's path.}

\item{platform}{Character. Platform of the input files, either `"10X"` or `"anndata"`. Default is `"10X"`.}

\item{use.names}{Logical. Whether to use the gene symbols rather the gene ids. WARNING: Only with 10X matrices. The gene names should be under the default /matrix/features/name directory of an 10X h5 file. Default is `TRUE`.}

\item{ensembl.to.symbol}{Logical. Whether to convert ENSEMBL IDs to gene symbols. Won't work if use.names is `TRUE`. Default is `FALSE`.}

\item{species}{Character. Species for gene symbol conversion (default `"human"`).}

\item{generate.metadata}{Logical. Whether to generate a basic metadata table with the procedence of the sample (default `FALSE`).}

\item{mc.cores}{Integer. Number of cores for `mclapply()` (default is the number of files).}
}
\value{
A named list of loaded/converted matrices (one per file).
}
\description{
For each file in `files.set` this function:
- constructs the full path using `files.dir`,
- if a directory with suffix `"_BP"` doesn't exist it reads the HDF5 matrix with
  `open_matrix_10x_hdf5()` and writes the matrix directory using `write_matrix_dir()`,
- loads the BP matrix directory with `open_matrix_dir()` and converts ENSEMBL IDs
  to gene symbols via `Azimuth:::ConvertEnsembleToSymbol()` if `ensembl.to.symbol` is `TRUE`, and
- generates a basic metadata table with the procedence of the sample if `generate.metadata` is `TRUE`.
}
\details{
Processing is parallelized with `parallel::mclapply()` on Unix-like systems.
On Windows, it falls back to `lapply()`.
}
\examples{
\dontrun{
BatchOpenH5(
  files = c("/path/to/h5ads/sample1.h5ad", "/path/to/h5ads/sample2.h5ad"),
  data.dir = "/path/to/data",
  mc.cores = 4
)
}

}
