% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateCDR.R
\name{CalculateCDR}
\alias{CalculateCDR}
\title{Calculate the cellular detection rate}
\usage{
CalculateCDR(SeuratObject)
}
\arguments{
\item{SeuratObject}{A Seurat object.}
}
\value{
The Seurat object with an additional "CDR" metadata slot with the calculated cdr.
}
\description{
Given a feature X cell table of raw counts, calculate the scaled cellular detection
rate for each cell. The cellular detection rate of a cell is a number between
0 and 1 representing the fraction of the total number of features measured
that have a count > 0 in that cell. It is useful as a covariate in
differential expression analyses. This function returns a vector of scaled CDRs.
}
\examples{
\dontrun{
SeuratObject <- CalculateCDR(SeuratObject)
}
}
