% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureScatterGradient.R
\name{FeatureScatterGradient}
\alias{FeatureScatterGradient}
\title{FeatureScatterGradient}
\usage{
FeatureScatterGradient(
  SeuratObject,
  feature1,
  feature2,
  gradient,
  upper.limit = NULL,
  lower.limit = 0,
  scale.colors = "viridis"
)
}
\arguments{
\item{SeuratObject}{The Seurat data.}

\item{feature1}{The feature to represent on x axis.}

\item{feature2}{The feature to represent on y axis.}

\item{gradient}{The feature which values will be used to set the color gradient.}

\item{upper.limit}{Upper limit of the gradient. When not specified, the gradient limits are automatically set.}

\item{lower.limit}{Lower limit of the gradient. Default is 0. Only applicable when upper.limit is specified.}

\item{scale.colors}{Color pallette to use in the gradient. Corresponds to the color palletes codes included in scale_gradient_viridis: "magma" (or "A"), "inferno" (or "B"), "plasma" (or "C"), "viridis" (or "D"), "cividis" (or "E"), "rocket" (or "F"), "mako" (or "G"), turbo" (or "H")}
}
\value{
A \code{ggplot2} object.
}
\description{
`FeatureScatterGradient()` adds the functionality over `Seurat::FeatureScatter()` to color the dots according a gradient determined by the value of another feature. This results very useful for visualizing QC parameters such as read count, feature count and % of mitochondrial genes simultaneously.
}
\examples{
\dontrun{
FeatureScatterGradient(SeuratObject, feature1="nCounts", feature2="nFeatures", gradient="percent.mt", upper.limit = 100, lower.limit = 0, scale.colors = "viridis")
}

}
