% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_functions.R
\name{seurat_recipe}
\alias{seurat_recipe}
\title{Apply a standard seurat workflow to a seurat object}
\usage{
seurat_recipe(
  seurat_object,
  assay = "RNA",
  nfeatures_vst = 1000,
  sample_column = NULL,
  remove_hvgs = FALSE,
  genes_to_remove = character(0),
  normalize_data = TRUE,
  npcs_PCA = 50,
  calcUMAP = TRUE,
  k.param = 20,
  findClusters = F,
  clusterRes = 1,
  key = "pca",
  seed = 123
)
}
\arguments{
\item{seurat_object}{seurat object}

\item{assay}{which assay}

\item{nfeatures_vst}{number of features for HVG detection}

\item{sample_column}{a column in meatdata with sample or batch ids. defaults to NULL which means feature detection will not be batch aware}

\item{remove_hvgs}{boolean. TRUE: use genes_to_remove to remove from HVGS}

\item{genes_to_remove}{a vector with genes to exclude}

\item{normalize_data}{normalize data with standard logNorm}

\item{npcs_PCA}{how many pcs to save}

\item{calcUMAP}{calc standard seurat UMAP}

\item{k.param}{k neighbors for UMAP and SNN}

\item{findClusters}{run neighbor and cluster detection with seurat standard}

\item{clusterRes}{resolution for clustering}

\item{key}{name/key for dimred (will name PCA like this and then use for all following steps)}

\item{seed}{random seed}
}
\value{
processed seurat_object
}
\description{
Apply a standard seurat workflow to a seurat object
}
